Add basic graph for states of averaged simulations
I now have functional code to create a data frame that is the average of multiple data frames for the number of people in each state for each day. I have also made the 3x3 graph for that data frame. The graph is not pretty yet and I am working on that next