Debug MLP Parameter NN and fix docs in data generator
Trained a model over 30,000 epochs with the MLPComponentsCV class. However, the data loaded in was already normalized. This was not known and thus, the data was divided by the population size once again. Causing the NN to not learn the parameters. Once this was found, proper documentation explaining that the data stored under 'Data/covasim/drums_data' for simulations generated using parallel computing is normalized prior to storage was added.